major update, changing all the hbm files over to the new schema
git-svn-id: https://svn.code.sf.net/p/specify/code/trunk/Specify@184 ce9bf4bf-376f-1848-a496-a1c9fe3b1a39

+++ src/edu/ku/brc/specify/datamodel/AttrsSettableGettable.java
@@ -20,14 +20,14 @@ public interface AttrsSettableGettable
     public void setIntValue(Integer intValue);
     
 
-    public Integer getFieldType();
+    public Short getFieldType();
     
-    public void setFieldType(Integer fieldType);
+    public void setFieldType(Short fieldType);
     
     
-    public Integer getUnit();
+    public Short getUnit();
     
-    public void setUnit(Integer unit);
+    public void setUnit(Short unit);
     
 
     public Date getTimestampCreated();
